project1/requirements.txt
pillow
numpy
PyMaxFlow
PySimpleGUI
matplotlib
scipy
opencv-python-headless